requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
httpx==0.28.1
gunicorn==23.0.0
torch==2.9.1
transformers==4.57.3
sentencepiece==0.2.1
requests==2.32.5
python-jose[cryptography]==3.5.0
python-dotenv==1.2.1
cryptography==46.0.3